[SPOT-286][TEST] 서비스 컨트롤러 분리에 따른 테스트 코드 분리
